Split triggers/embargo.py into focused submodules
Replace the 514-line flat module with a one-class-per-file package:
- embargo/propose.py  — SvcProposeEmbargoUseCase
- embargo/accept.py   — SvcAcceptEmbargoUseCase + SvcEvaluateEmbargoUseCase
- embargo/terminate.py — SvcTerminateEmbargoUseCase
- embargo/reject.py   — SvcRejectEmbargoUseCase
- embargo/revise.py   — SvcProposeEmbargoRevisionUseCase
- embargo/__init__.py — re-exports all public names for backward-compat

Mirror the split in tests: remove flat test_embargo.py and create
test/core/use_cases/triggers/embargo/ with per-submodule test files
and shared fixtures in conftest.py. All 12 existing tests are preserved.

Closes #969
