+++ test/core/use_cases/triggers/embargo/__init__.py
@@ -1,0 +1,1 @@
+"""Test package for embargo trigger use cases."""
